Refactor nodeInteractions.js and domainManagement.js into modular structure
- Decomposed nodeInteractions.js (1072 lines) into 5 focused modules
  - nodeSelection: Handles single and multi-node selection
  - nodeManipulation: Manages tags, deletion, and domain changes
  - nodeLinking: Handles link creation between nodes
  - ui: Controls selection panels and dialogs
  - events: Manages event bus communication

- Decomposed domainManagement.js (908 lines) into 6 focused modules
  - domainCollection: Gathers domains from graph data
  - domainModification: Handles creating, renaming, and deleting domains
  - domainSelection: Manages selecting nodes by domain
  - ui: Controls domain panel interface
  - colorManagement: Manages domain color assignments
  - pagination: Handles domain pagination

Each module now has a clear single responsibility, improving maintainability
while preserving the original API through re-exports.
